code refactoring part1
