+++ project/app.py
@@ -1,74 +1,105 @@
-from flask import Flask, render_template, redirect, url_for, flash, request, current_app, abort, jsonify,session
-from flask_wtf import CSRFProtect, FlaskForm
-from wtforms import StringField, PasswordField, SubmitField, TextAreaField, SelectField, ValidationError
-from flask_sqlalchemy import SQLAlchemy
-from flask_login import LoginManager, UserMixin, login_user, logout_user, login_required, current_user
-from werkzeug.security import generate_password_hash, check_password_hash
-from functools import wraps
+# --- Standard Library Imports ---
 import os
-from datetime import datetime, timedelta # Keep datetime for datetime.utcnow()
 import re
-import flask_socketio 
+import json
+import socket
+import base64
+from io import BytesIO
+from datetime import datetime, timedelta
+from functools import wraps
+
+# --- Flask Core Imports ---
+from flask import Flask, render_template, redirect, url_for, flash, request, current_app, abort, jsonify, session
+from flask_wtf import CSRFProtect, FlaskForm
+from flask_login import LoginManager, UserMixin, login_user, logout_user, login_required, current_user
 from flask_socketio import SocketIO, emit, join_room, leave_room, send
 
-# antonio: forms
-from forms import SignupForm,LoginForm,ReportForm,UpdateUserStatusForm,FriendRequestForm,UpdateReportStatusForm,Enable2FAForm,Disable2FAForm,RemovePassKeyForm
-
-# No clue but seems important
+# --- Database & ORM Imports ---
+from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.dialects.mysql import ENUM
 from sqlalchemy.exc import IntegrityError, SQLAlchemyError
 
-#parsing stufff for logs
-from parse_test import parse_modsec_audit_log,parse_error_log
-import socket
+# --- Forms & Validation Imports ---
+from wtforms import StringField, PasswordField, SubmitField, TextAreaField, SelectField, ValidationError
 
-# custom logs
-from user_actions import log_user_login_attempt, log_user_login_success, log_user_login_failure, log_user_logout
-
-#security stuff
+# --- Security & Authentication Imports ---
+from werkzeug.security import generate_password_hash, check_password_hash
 import pyotp
 import qrcode
-from io import BytesIO
-import base64
 from base64 import b64encode
 
-#web auth stuff
+# --- WebAuthn/Passkey Imports ---
 from fido2.server import Fido2Server
 from fido2.webauthn import PublicKeyCredentialRpEntity
-import json
 from fido2.client import ClientData
 from fido2.ctap2 import AttestationObject
 
+# --- Custom Module Imports ---
+# Models
+from models import (
+    db, User, Role, Permission, Event, Post, PostImage, 
+    Notification, Report, Chat, ChatParticipant, Message, 
+    Friendship, AdminAction, UserLog, ModSecLog, ErrorLog, 
+    WebAuthnCredential, user_role_assignments
+)
+
+# Filters
+from filters import (
+    apply_user_filters, apply_user_sorting, 
+    apply_report_filters, apply_user_log_filters
+)
+
+# Forms
+from forms import (
+    SignupForm, LoginForm, ReportForm, UpdateUserStatusForm,
+    FriendRequestForm, UpdateReportStatusForm, Enable2FAForm,
+    Disable2FAForm, RemovePassKeyForm
+)
+
+# Custom logging utilities
+from user_actions import (
+    log_user_login_attempt, log_user_login_success, 
+    log_user_login_failure, log_user_logout
+)
+
+# Log parsing utilities
+from parse_test import parse_modsec_audit_log, parse_error_log
+
 # --- Configuration ---
-# -- Flask app configuration --
+# Flask app configuration
 app = Flask(__name__)
 app.config['TEMPLATES_AUTO_RELOAD'] = True
 app.config['RECAPTCHA_PUBLIC_KEY'] = '6LdQNVsrAAAAAMp8AX4H_J4CwZ5OXVixltEf4RaC'
 app.config['RECAPTCHA_PRIVATE_KEY'] = '6LdQNVsrAAAAAMOmgh-7Tp-KAwQUQ6iIbi8_pRvM'
-socketio = SocketIO(app, cors_allowed_origins="https://localhost", message_queue='redis://redis:6379', logger=True, engineio_logger=True)   
 
-# -- Database configuration
+# Database configuration
 DB_USER = os.getenv('MYSQL_USER', 'flaskuser')
 DB_PASSWORD = os.getenv('MYSQL_PASSWORD', 'password')
 DB_NAME = os.getenv('MYSQL_DATABASE', 'flaskdb')
 DB_HOST = os.getenv('MYSQL_HOST', 'mysql')  # 'mysql' is the service name in docker-compose
 app.config['SQLALCHEMY_DATABASE_URI'] = f'mysql+pymysql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}/{DB_NAME}'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'a_very_secret_key_for_dev') # Change in production!
-db = SQLAlchemy(app)
+app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'a_very_secret_key_for_dev')  # Change in production!
+
+# Initialize database
+db.init_app(app)
+
+# Socket.IO configuration (MAKE DYNAMIC)
+socketio = SocketIO(app, cors_allowed_origins=[
+    "http://localhost",
+    "http://127.0.0.1",
+    "https://502a1f10a795.ngrok-free.app"
+])
 
 # --- Initialize Extensions ---
-# -- CSRF --
-app.config['SECRET_KEY'] = "your-very-secret-key"  # Set a secret key for CSRF protection
-csrf = CSRFProtect(app)  # Initialize CSRF protection
+# CSRF protection
+csrf = CSRFProtect(app)
 
-
-# -- Flask-Login --
+# Flask-Login
 login_manager = LoginManager(app)
-login_manager.init_app(app)
-login_manager.login_view = 'login' # Redirect to login if user not authenticated
+login_manager.login_view = 'login'  # Redirect to login if user not authenticated
 
-# -- Return container ID to templates --
+# --- Context Processors ---
 @app.context_processor
 def inject_container_id():
     return {"container_id": socket.gethostname()}
@@ -85,10 +116,6 @@ def load_user(user_id):
 
 # @login_required is already imported
 def user_required(f):
-    """
-    Decorator to restrict access to users who have ONLY the 'user' role.
-    Admins and editors (or anyone with more than just 'user' role) are denied.
-    """
     @wraps(f)
     @login_required
     def decorated_function(*args, **kwargs):
@@ -102,348 +129,13 @@ def user_required(f):
 
 # --- Admin Required Decorator ---
 def admin_required(f):
-    """
-    Decorator to restrict access to views to users with the 'admin' role.
-    Ensures user is logged in and has the 'admin' role.
-    """
     @wraps(f)
     @login_required # Ensure user is logged in first
     def decorated_function(*args, **kwargs):
         if not current_user.has_role('admin'):
-            flash('Access denied. You do not have administrator privileges.', 'danger')
-            # Abort with 403 Forbidden status code
             abort(403)
         return f(*args, **kwargs)
     return decorated_function
-
-# --- Models ---
-
-# **************************************
-# Association Tables (Many-to-Many Relationships)
-# **************************************
-
-user_role_assignments = db.Table('user_role_assignments',
-    db.Column('user_id', db.Integer, db.ForeignKey('users.user_id'), primary_key=True),
-    db.Column('role_id', db.Integer, db.ForeignKey('roles.role_id'), primary_key=True),
-    db.Column('assigned_at', db.DateTime, nullable=False, default=datetime.utcnow)
-)
-
-role_permissions = db.Table('role_permissions',
-    db.Column('role_id', db.Integer, db.ForeignKey('roles.role_id'), primary_key=True),
-    db.Column('permission_id', db.Integer, db.ForeignKey('permissions.permission_id'), primary_key=True)
-)
-
-
-# **************************************
-# 1. Users Table
-# **************************************
-class User(UserMixin, db.Model):
-    __tablename__ = 'users'
-    user_id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True, nullable=False)
-    phone_number = db.Column(db.String(8), unique=True, nullable=True)
-    password_hash = db.Column(db.String(255), nullable=False)
-    profile_pic_url = db.Column(db.String(255), nullable=True)
-    bio = db.Column(db.Text, nullable=True)
-    current_status = db.Column(db.String(50), nullable=False, default='offline')
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
-    last_active_at = db.Column(db.DateTime, nullable=True)
-    failed_login_attempts = db.Column(db.Integer, default=0)
-    lockout_until = db.Column(db.DateTime, nullable=True)
-    totp_secret = db.Column(db.String(32), nullable=True)
-    # Relationships
-    roles = db.relationship('Role', secondary=user_role_assignments,
-                            backref=db.backref('users', lazy='dynamic'), lazy='dynamic')
-    events = db.relationship('Event', backref='user', lazy=True)
-    posts = db.relationship('Post', backref='user', lazy=True)
-    notifications = db.relationship('Notification', backref='user', lazy=True)
-    submitted_reports = db.relationship(
-        'Report',
-        primaryjoin="User.user_id == Report.reporter_id", # Explicitly join User.user_id to Report.reporter_id
-        backref=db.backref('reporter_obj', lazy=True), # Use a unique backref name if needed, or let backref handle it
-        lazy=True
-    )
-
-    # A User can be reported in many reports
-    received_reports = db.relationship(
-        'Report',
-        primaryjoin="User.user_id == Report.reported_user_id", # Explicitly join User.user_id to Report.reported_user_id
-        backref=db.backref('reported_user_obj', lazy=True), # Use a unique backref name if needed
-        lazy=True
-    )
-    chat_participants = db.relationship('ChatParticipant', backref='user', lazy=True)
-    messages_sent = db.relationship('Message', foreign_keys='Message.sender_id', backref='sender', lazy=True)
-    admin_actions_performed = db.relationship('AdminAction', foreign_keys='AdminAction.admin_user_id', backref='admin_user', lazy=True)
-    admin_actions_targeted = db.relationship('AdminAction', foreign_keys='AdminAction.target_user_id', backref='target_user', lazy=True)
-    user_logs = db.relationship('UserLog', backref='user', lazy=True)
-
-    def set_password(self, password):
-        self.password_hash = generate_password_hash(password)
-
-    def check_password(self, password):
-        return check_password_hash(self.password_hash, password)
-
-    def get_id(self):
-        return str(self.user_id)
-
-    def has_role(self, role_name):
-        return any(role.role_name == role_name for role in self.roles)
-
-    def has_permission(self, permission_name):
-        for role in self.roles:
-            if any(perm.permission_name == permission_name for perm in role.permissions):
-                return True
-        return False
-
-    def __repr__(self):
-        return f"<User {self.username}>"
-
-
-# **************************************
-# 2. Roles and Permissions
-# **************************************
-class Role(db.Model):
-    __tablename__ = 'roles'
-    role_id = db.Column(db.Integer, primary_key=True)
-    role_name = db.Column(db.String(50), unique=True, nullable=False)
-    description = db.Column(db.Text, nullable=True)
-
-    permissions = db.relationship('Permission', secondary=role_permissions,
-                                  backref=db.backref('roles', lazy='dynamic'), lazy='dynamic')
-
-    def __repr__(self):
-        return f"<Role {self.role_name}>"
-
-class Permission(db.Model):
-    __tablename__ = 'permissions'
-    permission_id = db.Column(db.Integer, primary_key=True)
-    permission_name = db.Column(db.String(100), unique=True, nullable=False)
-    description = db.Column(db.Text, nullable=True)
-
-    def __repr__(self):
-        return f"<Permission {self.permission_name}>"
-
-
-# **************************************
-# 3. Events/Reminders
-# **************************************
-class Event(db.Model):
-    __tablename__ = 'events'
-    event_id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    title = db.Column(db.String(255), nullable=False)
-    description = db.Column(db.Text, nullable=True)
-    event_datetime = db.Column(db.DateTime, nullable=False)
-    location = db.Column(db.String(255), nullable=True)
-    is_reminder = db.Column(db.Boolean, nullable=False, default=False)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
-
-    def __repr__(self):
-        return f"<Event {self.title}>"
-
-# **************************************
-# 4. Posts
-# **************************************
-class Post(db.Model):
-    __tablename__ = 'posts'
-    post_id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    post_content = db.Column(db.Text, nullable=False)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
-
-    images = db.relationship('PostImage', backref='post', lazy=True)
-
-    def __repr__(self):
-        return f"<Post {self.post_id} by User {self.user_id}>"
-
-class PostImage(db.Model):
-    __tablename__ = 'post_images'
-    image_id = db.Column(db.Integer, primary_key=True)
-    post_id = db.Column(db.Integer, db.ForeignKey('posts.post_id'), nullable=False)
-    image_url = db.Column(db.String(255), nullable=False)
-    order_index = db.Column(db.Integer, nullable=True)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-
-    def __repr__(self):
-        return f"<PostImage {self.image_url[:20]}...>"
-
-
-# **************************************
-# 5. Notifications
-# **************************************
-class Notification(db.Model):
-    __tablename__ = 'notifications'
-    notification_id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    type = db.Column(ENUM('like', 'comment', 'friend_request', 'event_reminder', 'message', 'report_status', 'admin_override', name='notification_type'), nullable=False)
-    source_id = db.Column(db.Integer, nullable=True)
-    message = db.Column(db.String(255), nullable=False)
-    is_read = db.Column(db.Boolean, nullable=False, default=False)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-
-    def __repr__(self):
-        return f"<Notification {self.notification_id} for User:{self.user_id}>"
-
-
-# **************************************
-# 6. Customer Service
-# **************************************
-class Report(db.Model):
-    __tablename__ = 'reports'
-
-    report_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-
-    reporter_id = db.Column(db.Integer, db.ForeignKey('users.user_id', ondelete='SET NULL'), nullable=True)
-    reported_user_id = db.Column(db.Integer, db.ForeignKey('users.user_id', ondelete='CASCADE'), nullable=False)
-
-    report_type = db.Column(db.String(50), nullable=False)
-    description = db.Column(db.Text, nullable=False)
-    status = db.Column(db.String(20), nullable=False, default='open')
-    submitted_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    resolved_at = db.Column(db.DateTime, nullable=True)
-    admin_notes = db.Column(db.Text, nullable=True)
-
-    # Relationships:
-    # Add the overlaps parameter for the 'reporter' relationship
-    reporter = db.relationship('User', foreign_keys=[reporter_id],
-                               overlaps="reporter_obj,submitted_reports")
-
-    # This one already has the overlaps parameter from the previous fix
-    reported_user = db.relationship('User', foreign_keys=[reported_user_id],
-                                    overlaps="received_reports,reported_user_obj")
-
-    def __repr__(self):
-        return f"<Report {self.report_id} - Type: {self.report_type} - Status: {self.status}>"
-
-    def __repr__(self):
-        return f"<Report {self.report_id} - Type: {self.report_type} - Status: {self.status}>"
-
-
-# **************************************
-# 7. Messaging
-# **************************************
-class Chat(db.Model):
-    __tablename__ = 'chats'
-    chat_id = db.Column(db.Integer, primary_key=True)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-
-    participants = db.relationship('ChatParticipant', backref='chat', lazy=True)
-    messages = db.relationship('Message', backref='chat', lazy=True)
-
-    def __repr__(self):
-        return f"<Chat {self.chat_id}>"
-
-class ChatParticipant(db.Model):
-    __tablename__ = 'chat_participants'
-    chat_participant_id = db.Column(db.Integer, primary_key=True)
-    chat_id = db.Column(db.Integer, db.ForeignKey('chats.chat_id'), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-
-    __table_args__ = (db.UniqueConstraint('chat_id', 'user_id', name='_chat_user_uc'),)
-
-    def __repr__(self):
-        return f"<ChatParticipant Chat:{self.chat_id} User:{self.user_id}>"
-
-class Message(db.Model):
-    __tablename__ = 'messages'
-    message_id = db.Column(db.Integer, primary_key=True)
-    chat_id = db.Column(db.Integer, db.ForeignKey('chats.chat_id'), nullable=False)
-    sender_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    message_text = db.Column(db.Text, nullable=False)
-    sent_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    is_deleted_by_sender = db.Column(db.Boolean, nullable=False, default=False)
-    is_deleted_by_receiver = db.Column(db.Boolean, nullable=False, default=False)
-
-    def __repr__(self):
-        return f"<Message {self.message_id} in Chat:{self.chat_id} from User:{self.sender_id}>"
-
-
-# **************************************
-# 8. Friend System
-# **************************************
-class Friendship(db.Model):
-    __tablename__ = 'friendships'
-    friendship_id = db.Column(db.Integer, primary_key=True)
-    user_id1 = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    user_id2 = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    status = db.Column(ENUM('pending', 'accepted', 'blocked', name='friendship_status'), nullable=False, default='pending')
-    action_user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    updated_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
-
-    __table_args__ = (db.UniqueConstraint('user_id1', 'user_id2', name='_user1_user2_uc'),)
-
-    user1 = db.relationship('User', foreign_keys=[user_id1], backref=db.backref('friendships_as_user1', lazy='dynamic'))
-    user2 = db.relationship('User', foreign_keys=[user_id2], backref=db.backref('friendships_as_user2', lazy='dynamic'))
-    action_user = db.relationship('User', foreign_keys=[action_user_id], backref=db.backref('friendship_actions', lazy='dynamic'))
-
-    def __repr__(self):
-        return f"<Friendship {self.user_id1} - {self.user_id2} ({self.status})>"
-
-
-# **************************************
-# 9. Admin Panel
-# **************************************
-class AdminAction(db.Model):
-    __tablename__ = 'admin_actions'
-    action_id = db.Column(db.Integer, primary_key=True)
-    admin_user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    action_type = db.Column(db.String(100), nullable=False)
-    target_user_id = db.Column(db.Integer, db.ForeignKey('users.user_id', ondelete='SET NULL'), nullable=True)
-    target_entity_type = db.Column(db.String(50), nullable=True)
-    target_entity_id = db.Column(db.Integer, nullable=True)
-    details = db.Column(db.Text, nullable=True)
-    action_timestamp = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-
-    def __repr__(self):
-        return f"<AdminAction {self.action_type} by Admin:{self.admin_user_id}>"
-
-class UserLog(db.Model):
-    __tablename__ = 'user_logs'
-    log_id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    log_type = db.Column(db.String(100), nullable=False)
-    log_timestamp = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    details = db.Column(db.Text, nullable=True)
-
-class ModSecLog(db.Model): #actually in use
-    __tablename__ = 'ModSecLog'
-    id = db.Column(db.Integer, primary_key=True)
-    date = db.Column(db.String(20), nullable=False)
-    time = db.Column(db.String(20), nullable=False)
-    source = db.Column(db.String(50), nullable=False)
-    request = db.Column(db.Text, nullable=False)
-    response = db.Column(db.Text, nullable=False)
-    attack_detected = db.Column(db.Text, nullable=False)
-
-class ErrorLog(db.Model):
-    __tablename__ = 'ErrorLog'
-
-    id = db.Column(db.Integer, primary_key=True)  # Unique identifier for each log entry
-    date = db.Column(db.String(20), nullable=False)  # Date of the log (e.g., 2025/06/01)
-    time = db.Column(db.String(20), nullable=False)  # Time of the log (e.g., 12:40:46)
-    level = db.Column(db.Enum('notice', 'error', 'warning', 'critical'), nullable=False)  # Log level (e.g., notice, error)
-    message = db.Column(db.Text, nullable=False)  # Log message (e.g., limiting requests, excess: 3.295 by zone "api_limit")
-    client_ip = db.Column(db.String(50), nullable=False)  # Client IP address (e.g., 172.18.0.1)
-
-    def __repr__(self):
-        return f"<ErrorLog id={self.id} date={self.date} time={self.time} level={self.level} client_ip={self.client_ip}>"
-    
-
-# -- Webauth credentials -- 
-class WebAuthnCredential(db.Model):
-    __tablename__ = 'webauthn_credentials'
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
-    credential_id = db.Column(db.String(255), nullable=False, unique=True)
-    public_key = db.Column(db.Text, nullable=False)
-    sign_count = db.Column(db.Integer, nullable=False)
-    nickname = db.Column(db.String(100))
-    added_at = db.Column(db.DateTime, default=datetime.utcnow)
-    user = db.relationship('User', backref='webauthn_credentials')
 
 # --- Flask-Login User Loader ---
 @login_manager.user_loader
@@ -451,40 +143,41 @@ def load_user(user_id):
     return db.session.get(User, int(user_id))
 
 # --- RBAC Decorators ---
-def role_required(role_name):
-    def decorator(f):
-        @wraps(f)
-        @login_required
-        def decorated_function(*args, **kwargs):
-            if not current_user.has_role(role_name):
-                flash(f'You do not have the required role: {role_name}.', 'danger')
-                return redirect(url_for('dashboard'))
-            return f(*args, **kwargs)
-        return decorated_function
-    return decorator
+#def role_required(role_name):
+    #def decorator(f):
+        #@wraps(f)
+        #@login_required
+        #def decorated_function(*args, **kwargs):
+            #if not current_user.has_role(role_name):
+                #return redirect(url_for('dashboard'))
+            #return f(*args, **kwargs)
+        #return decorated_function
+    #return decorator
 
-def permission_required(permission_name):
-    def decorator(f):
-        @wraps(f)
-        @login_required
-        def decorated_function(*args, **kwargs):
-            if not current_user.has_permission(permission_name):
-                flash(f'You do not have the required permission: {permission_name}.', 'danger')
-                return redirect(url_for('dashboard'))
-            return f(*args, **kwargs)
-        return decorated_function
-    return decorator
+#def permission_required(permission_name):
+    #ef decorator(f):
+        #@wraps(f)
+        #@login_required
+        #def decorated_function(*args, **kwargs):
+            #if not current_user.has_permission(permission_name):
+                #flash(f'You do not have the required permission: {permission_name}.', 'danger')
+                #return redirect(url_for('dashboard'))
+            #return f(*args, **kwargs)
+        #return decorated_function
+    #return decorator
 
 
 # -- Fido2 WebAuthn Server Setup --
+# -- For passkey
+
 def get_fido2_server():
     from flask import request
     rp_id = request.host.split(':')[0]
     rp_name = "SimpleBook"
     rp = PublicKeyCredentialRpEntity(rp_id, rp_name)
     return Fido2Server(rp)
+
 def b64encode_all(obj):
-    import base64
     if isinstance(obj, dict):
         return {k: b64encode_all(v) for k, v in obj.items()}
     elif isinstance(obj, list):
@@ -493,8 +186,6 @@ def b64encode_all(obj):
         return base64.b64encode(obj).decode('utf-8')
     else:
         return obj
-# --- Routes ---
-# index = home
 
 
 # --- login,signup,home ---
@@ -537,10 +228,12 @@ def login():
                     user.failed_login_attempts = 0
                     user.lockout_until = None
                     db.session.commit()
+
                     login_user(user)
                     user.current_status = 'online'
                     user.last_active_at = datetime.utcnow()
                     db.session.commit()
+
                     # --- Log user login success ---
                     log_user_login_success(user.user_id, details="User logged in successfully.")
                     next_page = request.args.get('next')
@@ -558,20 +251,16 @@ def login():
                         log_user_login_failure(user.user_id, details="User failed login attempt.")
                     db.session.commit()
                 else:
-                    # Optionally, log attempts for non-existent users (not recommended for privacy)
-                    flash('Invalid username or password.', 'danger')
+                    pass
     return render_template('UserLogin.html', form=form)
 
 @app.route('/logout')
 @login_required
 def logout():
     if current_user.is_authenticated:
-        # --- NEW: Update user status on logout ---
         current_user.current_status = 'offline'
         db.session.commit()
-    # --- NEW: Log user logout action ---
     log_user_logout(current_user.user_id, details="User logged out.")
-    # --- END NEW ---
     logout_user()
     return redirect(url_for('login'))
 
@@ -597,45 +286,39 @@ def signup():
             password_hash=generate_password_hash(password),
             created_at=datetime.utcnow(),
             updated_at=datetime.utcnow(),
-            
+            current_status='online'
         )
         db.session.add(new_user)
         db.session.commit()
 
+        # Assign 'user' role
         default_role = Role.query.filter_by(role_name='user').first()
-        if default_role:
-            new_user.roles.append(default_role)
-            new_user.current_status = 'online'
-            db.session.commit()
-            login_user(new_user)
-            return redirect(url_for('home'))
-        else:
-            return redirect(url_for('login'))
+        new_user.roles.append(default_role)
+        db.session.commit()
+        
+        login_user(new_user)
+        return redirect(url_for('home'))
 
     return render_template('UserSignup.html', form=form)
 
 @app.route('/verify_2fa', methods=['GET', 'POST'])
 def verify_2fa():
+    
     if 'pending_2fa_user_id' not in session:
-        flash('Session expired. Please log in again.', 'danger')
         return redirect(url_for('login'))
 
     user = User.query.get(session['pending_2fa_user_id'])
     if not user or not user.totp_secret:
-        flash('Invalid session or 2FA not enabled.', 'danger')
         return redirect(url_for('login'))
 
-    form = Enable2FAForm()  # Reuse your 2FA form
+    form = Enable2FAForm()
     if form.validate_on_submit():
         code = form.totp_code.data
         totp = pyotp.TOTP(user.totp_secret)
         if totp.verify(code):
             login_user(user)
             session.pop('pending_2fa_user_id', None)
-            flash('Logged in successfully!', 'success')
             return redirect(url_for('home'))
-        else:
-            flash('Invalid 2FA code.', 'danger')
     return render_template('UserVerify2FA.html', form=form)
 
 
@@ -654,17 +337,17 @@ def account_security():
 @user_required
 def enable_2fa():
     form = Enable2FAForm()
+
+    # -- Already enabled --
     if current_user.totp_secret:
-        flash('2FA is already enabled.', 'info')
         return redirect(url_for('account_security'))
 
-    # Only generate a new secret if not already in session
     if 'pending_totp_secret' not in session:
         session['pending_totp_secret'] = pyotp.random_base32()
     totp_secret = session['pending_totp_secret']
 
     otp_uri = pyotp.TOTP(totp_secret).provisioning_uri(
-        name=current_user.username, issuer_name="YourApp"
+        name=current_user.username, issuer_name="SimpleBook"
     )
     img = qrcode.make(otp_uri)
     buf = BytesIO()
@@ -678,11 +361,7 @@ def enable_2fa():
             current_user.totp_secret = totp_secret
             db.session.commit()
             session.pop('pending_totp_secret', None)  # Remove from session
-            flash('2FA enabled successfully!', 'success')
             return redirect(url_for('account_security'))
-        else:
-            flash('Invalid code. Please try again.', 'danger')
-
     return render_template('UserEnable2FA.html', qr_b64=qr_b64, secret=totp_secret, form=form)
 
 @app.route('/disable_2fa', methods=['POST'])
@@ -692,9 +371,7 @@ def disable_2fa():
     if form.validate_on_submit():
         current_user.totp_secret = None
         db.session.commit()
-        flash('2FA has been disabled.', 'info')
         return redirect(url_for('account_security'))
-    # If not valid or GET (shouldn't happen for POST-only), re-render the page
     return render_template('UserDisable2FA.html', form=form)
 
 # -- User passkey management --
@@ -722,10 +399,8 @@ def passkey_begin_register(): #stable!
         encoded = b64encode_all(registration_data)
         # --- FIX: Wrap in 'publicKey' ---
         return jsonify({"publicKey": encoded})
-    except Exception as e:
-        import traceback
-        print(traceback.format_exc())
-        return jsonify({"error": str(e)}), 400
+    except:
+        return redirect(url_for('account_security'))
 
 @app.route('/passkey/finish_register', methods=['POST'])
 @csrf.exempt
@@ -800,29 +475,17 @@ def report_user():
         try:
             db.session.add(new_report)
             db.session.commit()
-            flash('Your report has been submitted successfully. Thank you for helping us keep the community safe!', 'success')
             # Redirect to the profile of the user who was reported, or a confirmation page
             return redirect(url_for('report_confirmation', reported_username=reported_user.username))
-        except IntegrityError as e:
+        except:
+            #error or smth
             db.session.rollback()
-            current_app.logger.error(f"IntegrityError submitting report: {e}", exc_info=True)
-            flash('Could not submit report due to a data conflict.', 'danger')
-        except SQLAlchemyError as e:
-            db.session.rollback()
-            current_app.logger.error(f"Database error submitting report: {e}", exc_info=True)
-            flash('An unexpected database error occurred. Please try again later.', 'danger')
-        except Exception as e:
-            db.session.rollback() # Rollback in case of any unhandled exception during DB ops
-            current_app.logger.error(f"Unhandled exception submitting report: {e}", exc_info=True)
-            flash('An unknown error occurred while submitting your report. Please try again.', 'danger')
 
-    # For GET request or if validation fails, render the form
     return render_template('UserReport.html', form=form)
 
 @app.route('/report_confirmation')
 @user_required
 def report_confirmation():
-    # You can get the reported_username from the query parameters if passed
     reported_username = request.args.get('reported_username', 'the user')
     return render_template('UserReportConfirmed.html', reported_username=reported_username)
 
@@ -1213,10 +876,9 @@ def chat_history(friend_id):
 
 # --- Admin Routes ---
 @app.route('/users_dashboard')
-@admin_required # Protect this route with the admin_required decorator
+@admin_required
 def manage_users():
     # Get query parameters
-    
     search_query = request.args.get('search', '').strip()
     sort_by = request.args.get('sort_by', 'id')  # Default sort by ID
     order = request.args.get('order', 'asc')  # Default order is ascending
@@ -1224,47 +886,11 @@ def manage_users():
     # Base query
     query = User.query
 
-    # Apply search filters with enhanced validation
-    if search_query:
-        filters = search_query.split(',')
-        for filter_item in filters:
-            filter_item = filter_item.strip()  # Remove extra spaces
-            filter_item = re.sub(r'[^\w=]', '', filter_item)  # Remove special characters except '='
-
-            if 'id=' in filter_item:
-                try:
-                    user_id = int(filter_item.split('id=')[1])
-                    query = query.filter(User.user_id == user_id)
-                except ValueError:
-                    flash("Invalid ID format. ID must be a number.", "danger")
-            elif 'username=' in filter_item:
-                username = filter_item.split('username=')[1].strip()
-                if re.match(r'^[a-zA-Z0-9_]+$', username):  # Allow alphanumeric and underscores
-                    query = query.filter(User.username.ilike(f"%{username}%"))
-                else:
-                    flash("Invalid username format. Username must be alphanumeric.", "danger")
-            elif 'phone=' in filter_item:
-                phone = filter_item.split('phone=')[1].strip()
-                if re.match(r'^\d+$', phone):  # Ensure phone contains only digits
-                    query = query.filter(User.phone_number.ilike(f"%{phone}%"))
-                else:
-                    flash("Invalid phone format. Phone must contain only digits.", "danger")
-            elif 'status=' in filter_item:
-                status = filter_item.split('status=')[1].strip().lower()
-                if status in ['online', 'offline']:  # Ensure status is valid
-                    query = query.filter(User.current_status.ilike(f"%{status}%"))
-                else:
-                    flash("Invalid status format. Status must be 'online' or 'offline'.", "danger")
-            else:
-                flash("Invalid query format. Please use id=, username=, phone=, or status=.", "danger")
-
-    # Apply sorting
-    if sort_by == 'username':
-        query = query.order_by(User.username.asc() if order == 'asc' else User.username.desc())
-    elif sort_by == 'registration_date':
-        query = query.order_by(User.created_at.asc() if order == 'asc' else User.created_at.desc())
-    else:  # Default sort by ID
-        query = query.order_by(User.user_id.asc() if order == 'asc' else User.user_id.desc())
+    # Apply filters using the separate function
+    query = apply_user_filters(query, search_query)
+    
+    # Apply sorting using the separate function
+    query = apply_user_sorting(query, sort_by, order)
 
     users = query.all()
 
@@ -1283,7 +909,6 @@ def manage_users():
         search_query=search_query,
         form=UpdateUserStatusForm()
     )
-
 @app.route('/manage_user/<int:user_id>', methods=['GET', 'POST'])
 @admin_required
 def manage_user(user_id):
@@ -1310,39 +935,16 @@ def manage_user(user_id):
 def manage_reports():
     # Get query parameters for filtering and sorting
     search_query = request.args.get('search', '').strip()
-    sort_by = request.args.get('sort_by', 'submitted_at')  # Default sort by submission date
-    order = request.args.get('order', 'desc')  # Default order is descending
+    sort_by = request.args.get('sort_by', 'submitted_at')
+    order = request.args.get('order', 'desc')
 
     # Base query
     query = Report.query
 
-    # Apply search filters
-    if search_query:
-        filters = search_query.split(',')
-        for filter_item in filters:
-            filter_item = filter_item.strip()
-            if 'report_id=' in filter_item:
-                try:
-                    report_id = int(filter_item.split('report_id=')[1])
-                    query = query.filter(Report.report_id == report_id)
-                except ValueError:
-                    flash("Invalid report ID format. ID must be a number.", "danger")
-            elif 'status=' in filter_item:
-                status = filter_item.split('status=')[1].strip().lower()
-                if status in ['open', 'in_review', 'action_taken', 'rejected']:
-                    query = query.filter(Report.status.ilike(f"%{status}%"))
-                else:
-                    flash("Invalid status format. Status must be 'open', 'in_review', 'action_taken', or 'rejected'.", "danger")
-            elif 'report_type=' in filter_item:
-                report_type = filter_item.split('report_type=')[1].strip().lower()
-                if report_type in ['spam', 'harassment', 'impersonation', 'inappropriate_content', 'fraud', 'other']:
-                    query = query.filter(Report.report_type.ilike(f"%{report_type}%"))
-                else:
-                    flash("Invalid report type format.", "danger")
-            else:
-                flash("Invalid query format. Please use report_id=, status=, or report_type=.", "danger")
+    # Apply filters using the separate function
+    query = apply_report_filters(query, search_query)
 
-    # Apply sorting
+    # Apply sorting (you could move this to filters.py too)
     if sort_by == 'submitted_at':
         query = query.order_by(Report.submitted_at.asc() if order == 'asc' else Report.submitted_at.desc())
     elif sort_by == 'resolved_at':
@@ -1369,6 +971,7 @@ def manage_reports():
         action_taken_reports=action_taken_reports,
         rejected_reports=rejected_reports
     )
+
 
 @app.route('/manage_report/<int:report_id>', methods=['GET', 'POST'])
 @admin_required
@@ -1584,33 +1187,10 @@ def admin_user_actions():
 
     query = UserLog.query
 
-    # Filtering
-    if search_query:
-        filters = search_query.split(',')
-        for filter_item in filters:
-            filter_item = filter_item.strip()
-            if 'id=' in filter_item:
-                try:
-                    log_id = int(filter_item.split('id=')[1])
-                    query = query.filter(UserLog.log_id == log_id)
-                except ValueError:
-                    flash("Invalid ID format. ID must be a number.", "danger")
-            elif 'user_id=' in filter_item:
-                try:
-                    user_id = int(filter_item.split('user_id=')[1])
-                    query = query.filter(UserLog.user_id == user_id)
-                except ValueError:
-                    flash("Invalid user ID format.", "danger")
-            elif 'log_type=' in filter_item:
-                log_type = filter_item.split('log_type=')[1].strip()
-                query = query.filter(UserLog.log_type.ilike(f"%{log_type}%"))
-            elif 'date=' in filter_item:
-                date = filter_item.split('date=')[1].strip()
-                query = query.filter(UserLog.log_timestamp.ilike(f"%{date}%"))
-            else:
-                flash("Invalid query format. Please use id=, user_id=, log_type=, or date=.", "danger")
+    # Apply filters using the separate function
+    query = apply_user_log_filters(query, search_query)
 
-    # Only allow sorting by id or log_timestamp
+    # Apply sorting
     if sort_by == 'log_timestamp':
         query = query.order_by(UserLog.log_timestamp.asc() if order == 'asc' else UserLog.log_timestamp.desc())
     else:  # Default sort by ID
@@ -1633,7 +1213,6 @@ def admin_user_actions():
         order=order,
         search_query=search_query
     )
-
 
 # --- Error Handlers ---
 @app.errorhandler(404)
